requirements.txt
absl-py==2.1.0
ai-edge-litert==1.0.1
ai-edge-model-explorer==0.1.14
ai-edge-model-explorer-adapter==0.1.5
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.10.0
gast==0.4.0
google-auth==2.36.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.68.0
h5py==3.12.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
onnx==1.17.0
onnx-graphsurgeon==0.5.2
onnx-simplifier==0.4.36
onnx2tf==1.20.0
onnxruntime==1.19.2
onnxslim==0.1.31
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
portpicker==1.6.0
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sng4onnx==1.0.1
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.18.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.16.0
tflite-support==0.1.0a1
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.16.0
zipp==3.21.0
